requirements.txt
databases==0.7.0
SQLAlchemy==1.4.49
fastapi==0.100.0
starlette==0.27.0
jose==1.0.0
pydantic==1.10.11
python-dotenv==1.0.0
passlib==1.7.4
uvicorn==0.22.0
psycopg2==2.9.6
requests==2.31.0
rocketry==2.5.1